creating routers and the db connection for local
