+++ workflow_engine/requirements.txt
@@ -3,3 +3,4 @@ psycopg2-binary
 alembic
 fastapi
 uvicorn[standard]
+python-dotenv
