+++ workflow_engine/workflow/dependencies.py
@@ -1,0 +1,9 @@
+from typing import Generator
+from workflow.db.database import SessionLocal
+
+def get_db() -> Generator:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
